% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apa.R
\name{anova_apa}
\alias{anova_apa}
\title{Report ANOVA in APA style}
\usage{
anova_apa(x, sph_corr = "greenhouse-geisser", es = "petasq",
  format = c("text", "markdown", "rmarkdown", "html", "latex", "docx"),
  info = FALSE, print = TRUE)
}
\arguments{
\item{x}{a call to \code{ez::ezANOVA}.}

\item{sph_corr}{character string indicating the method used for correction if
sphericity is violated (only applies to repeated-measures and mixed design
ANOVA). Can be one of \code{"greenhouse-geisser"} (default),
\code{"huynh-feldt"} or \code{"none"} (you may also use the abbreviations
\code{"gg"} or \code{"hf"}).}

\item{es}{character string indicating the effect size to show in the output,
one of \code{"petasq"} or \code{"getasq"}.}

\item{format}{character string specifying the output format. One of
\code{"default"}, \code{"text"}, \code{"markdown"}, \code{"rmarkdown"},
\code{html}, \code{"latex"} or \code{"docx"}.}

\item{info}{logical indicating whether to print a message on the used test
(default is \code{FALSE})}

\item{print}{logical indicating wheter to print the formatted output via
\code{cat} (\code{TRUE}, default) or return as a data frame.}
}
\description{
Report ANOVA in APA style
}
\examples{
\dontrun{
library(dplyr)
library(tidyr)

# Convert data from wide format to long format
data <- height \%>\%
  select(id, group, anx_lvl1:anx_lvl3) \%>\%
  gather(key = "level", value = "anxiety", anx_lvl1:anx_lvl3)

# Use ez package
library(ez)
ezANOVA(data, dv = anxiety, wid = id, within = level, between = group,
        detailed = TRUE) \%>\%
  anova_apa()

# Use afex package
library(afex)
aov_ez(id = "id", dv = "anxiety", data = data, between = "group",
       within = "level") \%>\%
  anova_apa()
}
}

