% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apa.R
\name{chisq_apa}
\alias{chisq_apa}
\title{Report Chi-squared test in APA style}
\usage{
chisq_apa(x, print_n = FALSE, format = c("default", "text", "markdown",
  "rmarkdown", "html", "latex", "docx"), info = FALSE)
}
\arguments{
\item{x}{a call to \code{chisq.test}}

\item{print_n}{logical indicating whether to show sample size in text}

\item{format}{character string specifying the output format. One of
\code{"default"}, \code{"text"}, \code{"markdown"}, \code{"rmarkdown"},
\code{html}, \code{"latex"} or \code{"docx"}.}

\item{info}{logical indicating whether to print a message on the used test
(default is \code{FALSE})}
}
\description{
Report Chi-squared test in APA style
}
\examples{
chisq_apa(chisq.test(height$group, height$sex))
}

