% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apa.R
\name{cor_apa}
\alias{cor_apa}
\title{Report Correlation in APA style}
\usage{
cor_apa(x, format = c("default", "text", "markdown", "rmarkdown", "html",
  "latex", "docx"), info = FALSE)
}
\arguments{
\item{x}{a call to \code{cor.test}}

\item{format}{character string specifying the output format. One of
\code{"default"}, \code{"text"}, \code{"markdown"}, \code{"rmarkdown"},
\code{html}, \code{"latex"} or \code{"docx"}.}

\item{info}{logical indicating whether to print a message on the used test
(default is \code{FALSE})}
}
\description{
Report Correlation in APA style
}
\examples{
cor_apa(cor.test(~ anx_lvl1 + anx_lvl2, height))
}

