% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cor_table.R
\name{cor_table}
\alias{cor_table}
\title{Correlation table}
\usage{
cor_table(data, adjust = NULL, labels = names(data), format = c("text",
  "html", "latex", "markdown", "rmarkdown"), part = c("lower", "upper"))
}
\arguments{
\item{data}{a data frame}

\item{adjust}{character string indicating the method used for adjusting
p-values for muliple comparisons. See \link{p.adjust.methods} for available
methods. If is \code{NULL}, no adjustment is applied (default).}

\item{labels}{character vector indicating the labels used for header row and
first column}

\item{format}{character string indicating the output format, one of
\code{"text"}, \code{"html"} or \code{"latex"}. For HTML, the table is
displayed as a htmlwidget. For LaTeX, the \code{tabularx} environment is
used, which requires the tabularx package.}

\item{part}{character string indicating which part of the correlation matrix
should be printed, can be either \code{"lower"} or \code{"upper"}.}
}
\description{
Correlation table
}
\examples{
cor_table(height[4:9])
}
\seealso{
\link{cor}
}

