% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplot.R
\name{fplot}
\alias{fplot}
\title{fplot}
\usage{
fplot(.data, formula, geom = c("bar", "line", "boxplot"), error = c("se",
  "sd", "moe"), ...)
}
\arguments{
\item{.data}{a data frame containing the variables in the formula
\code{formula}.}

\item{formula}{a formula in the form \code{lhs ~ rhs} where \code{lhs} is a
numeric variable (the dependent variable) and \code{rhs} one or more
factors with two or more levels giving the corresponding groups.}

\item{geom}{character string indicating the ggplot2 geom used. One of
\code{"bar"} (default), \code{"line"} and \code{"boxplot"}.}

\item{error}{character string indicating the function used for calculating
the error bars. One of \code{"se"} (default) \code{"sd"} or \code{"moe"}.
Ignored if \code{geom} is \code{"boxplot"}.}

\item{...}{Further arguments passed to methods. E.g. \code{quantile} if
\code{error} is \code{"moe"} to indicate the quantile for the confidence
interval.}
}
\description{
Convenient plotting of means and error bars or boxplots of factorial
data.
}
\examples{
fplot(hquest, trait_anx ~ group + gender)

# Standard deviation instead of standard error
fplot(hquest, trait_anx ~ group + gender, error = "sd")

# 95\% confidence intervals (moe = margin of error)
fplot(hquest, trait_anx ~ group + gender, error = "moe")

# Line graph
fplot(hquest, trait_anx ~ group + gender, geom = "line")

# Boxplot
fplot(hquest, trait_anx ~ group + gender, geom = "boxplot")

}

