% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hquest.R
\docType{data}
\name{hquest}
\alias{hquest}
\title{Simulated questionnaire data}
\format{A data frame with 10 rows and 7 variables:
\describe{
  \item{id}{Subject identifier}
  \item{group}{Between subject factor (height anxious vs. control)}
  \item{gender}{Gender of the participants}
  \item{age}{Age of the participants}
  \item{trait_anx}{Trait anxiety (0-100)}
  \item{acrophobia}{Height anxiety (0-100)}
  \item{sens_seek}{Sensation seeking (0-100)}
}}
\usage{
hquest
}
\description{
Simulated questionnaire data from a height anxiety experiment.
}
\keyword{datasets}

