% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_cor.R
\name{mean_cor}
\alias{mean_cor}
\title{Mean correlation using Fisher-Z-transformation}
\usage{
mean_cor(...)
}
\arguments{
\item{...}{correlation coefficients to average (either as single numeric,
a numeric vector or a list)}
}
\description{
Mean correlation using Fisher-Z-transformation
}
\examples{
mean_cor(.45, .33, .36)

}

