% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ci.R
\name{moe}
\alias{moe}
\title{Margin of Error}
\usage{
moe(x, quantile = 0.975, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an \R object which is coercible to one by
\code{as.vector(x, "numeric")}}

\item{quantile}{numeric, quantile of the t-distribution. Default is
\code{.975} for half the width of a 95\% confidence interval.}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Margin of Error
}

