% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/effect_size.R
\name{petasq}
\alias{petasq}
\title{Partial Eta Squared}
\usage{
petasq(x, effect)
}
\arguments{
\item{x}{a call to \code{aov} or \code{ez::ezANOVA}}

\item{effect}{character string indicating the effect of interest}
}
\description{
Partial Eta Squared
}

