% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/routlier.R
\name{routlier}
\alias{routlier}
\title{Remove or flag outliers in a data frame}
\usage{
routlier(.data, ..., action = c("na", "flag", "remove"), z = c(-3, 3),
  absolute = NULL, info = FALSE)
}
\arguments{
\item{.data}{a data frame}

\item{...}{names of the columns to check for outliers}

\item{action}{character string indicating how to deal with outliers, can be
one of \code{"na"} (outliers are turned into \code{NA}; default),
\code{"remove"} (rows containing outliers are removed from the data frame)
or \code{"flag"} (a new logical variable "outlier" is added to the data
frame indicating which rows contain outliers).}

\item{z}{numeric vector of length two indicating the lower and upper
z-score limits for outlier detection. If only one side should be scanned
use e.g. \code{c(NULL, 3)}.}

\item{absolute}{numeric vector of length two indicating the absolute lower
and upper limits for outlier detection. If only one side should be scanned
use e.g. \code{c(10, NULL)}.}

\item{info}{logical indicating whether to print the number of outliers
  detected to the console.}
}
\description{
Remove or flag outliers in a data frame
}

