% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard Error of the Mean}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an \R object which is coercible to one by
\code{as.vector(x, "numeric")}}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
This function computes the standard error of the mean of the values in
\code{x} using the formula \code{se = sd(x) / sqrt(length(x))}. If
\code{na.rm} is \code{TRUE} then missing values are removed before
computation proceeds.
}
\seealso{
\link{sd}
}

