% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_table.R
\name{stats_table}
\alias{stats_table}
\title{Statistical table}
\usage{
stats_table(.data, iv, dvs, funs = c("mean", "sd"), paired = FALSE, id,
  sig = FALSE, format = c("text", "html", "latex"), ...)
}
\arguments{
\item{.data}{a data frame}

\item{iv}{name of the independent variable}

\item{dvs}{names of the dependent variables}

\item{funs}{character vector with function names indicating the parameters to
calculate (default: \code{c("mean", "sd")}).}

\item{paired}{logical indicating whether you have paired data (\code{iv} is a
within factor) or non-paired data (\code{iv} is a between factor).}

\item{id}{name or character string indicating the subject identifier column
in \code{.data}. Only required if the independent variable has more than
two levels.}

\item{sig}{logical indicating whether to show significance with symbols in
the last column}

\item{format}{character string indicating the output format, one of
\code{"text"}, \code{"html"} or \code{"latex"}. For HTML, the table is
displayed as a htmlwidget. For LaTeX, the \code{tabularx} environment is
used, which requires the tabularx package. The utilized \code{\\cmidrule}
command requires either the booktabs or ctable package.}

\item{...}{Further arguments passed to functions. If the independent variable
has two levels, then \code{...} can contain arguments passed to
\link{t_test}, e.g. \code{var.equal = TRUE}.}
}
\description{
Compiles a table with descriptive and inferential statistics to compare group
characteristics, e.g. control variables of a treatment group and a control
group.
}
\examples{
stats_table(hquest, iv = group, dvs = age:sens_seek)
stats_table(hquest, iv = group, dvs = age:sens_seek, sig = TRUE,
            format = "html", var.equal = TRUE)
}

