% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa.R
\name{anova_apa}
\alias{anova_apa}
\title{Report ANOVA in APA style}
\usage{
anova_apa(x, effect = NULL, sph_corr = c("greenhouse-geisser", "gg",
  "huynh-feldt", "hf", "none"), es = c("petasq", "pes", "getasq", "ges"),
  format = c("text", "markdown", "rmarkdown", "html", "latex", "docx",
  "plotmath"), info = FALSE, print = TRUE)
}
\arguments{
\item{x}{a call to \code{ez::ezANOVA} or \code{afex::afex_ez},
\code{afex::afex_car} or \code{afex::afex_4}}

\item{effect}{character string indicating the name of the effect to display.
If is \code{NULL}, all effects are reported (default).}

\item{sph_corr}{character string indicating the method used for correction if
sphericity is violated (only applies to repeated-measures and mixed design
ANOVA). Can be one of \code{"greenhouse-geisser"} (default),
\code{"huynh-feldt"} or \code{"none"} (you may also use the abbreviations
\code{"gg"} or \code{"hf"}).}

\item{es}{character string indicating the effect size to display in the
output, one of \code{"petasq"} (partial eta squared) or \code{"getasq"}
(generalized eta squared) (you may also use the abbreviations \code{"pes"}
or \code{"ges"}).}

\item{format}{character string specifying the output format. One of
\code{"text"}, \code{"markdown"}, \code{"rmarkdown"}, \code{html},
\code{"latex"}, \code{"docx"} or \code{"plotmath"}.}

\item{info}{logical indicating whether to print a message on the used test
(default is \code{FALSE})}

\item{print}{logical indicating wheter to print the formatted output via
\code{cat} (\code{TRUE}, default) or return as a data frame.}
}
\description{
Report ANOVA in APA style
}
\examples{
\dontrun{
library(magrittr)

# Using the ez package
library(ez)
ezANOVA(height, dv = anxiety, wid = id, within = level, between = group,
        detailed = TRUE) \%>\%
  anova_apa()

# Using the afex package
library(afex)
aov_ez(id = "id", dv = "anxiety", data = height, between = "group",
       within = "level") \%>\%
  anova_apa()
}

}

