% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apa.R
\name{anova_apa}
\alias{anova_apa}
\title{Report ANOVA in APA style}
\usage{
anova_apa(x, sph_corr = "greenhouse-geisser", es = "petasq",
  format = c("default", "text", "markdown", "rmarkdown", "html", "latex",
  "docx"), info = FALSE)
}
\arguments{
\item{x}{a call to \code{ez::ezANOVA}.}

\item{sph_corr}{character string indicating the method used for correction if
sphericity is violated (only applies to repeated-measures and mixed design
ANOVA). Can be one of \code{"greenhouse-geisser"} (default),
\code{"huynh-feldt"} or \code{"none"} (you may also use the abbreviations
\code{"gg"} or \code{"hf"}).}

\item{es}{character string indicating the effect size to show in the output,
one of \code{"petasq"} or \code{"getasq"}.}

\item{format}{character string specifying the output format. One of
\code{"default"}, \code{"text"}, \code{"markdown"}, \code{"rmarkdown"},
\code{html}, \code{"latex"} or \code{"docx"}.}

\item{info}{logical indicating whether to print a message on the used test
(default is \code{FALSE})}
}
\description{
Report ANOVA in APA style
}

