% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apa.R
\name{chisq_apa}
\alias{chisq_apa}
\title{Report Chi-squared test in APA style}
\usage{
chisq_apa(x, print_n = FALSE, format = "default", info = FALSE)
}
\arguments{
\item{x}{a call to \code{chisq.test}}

\item{print_n}{logical indicating whether to show sample size in text}

\item{format}{character specifying the output format, one of
\code{"default"}, \code{"text"}, \code{"latex"}, \code{"markdown"},
\code{"rmarkdown"} or \code{"docx"}.}

\item{info}{logical indicating whether to print a message on the used test
  (default is \code{FALSE})}
}
\description{
Report Chi-squared test in APA style
}

