% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/effect_size.R
\name{cohens_d}
\alias{cohens_d}
\alias{cohens_d.data.frame}
\alias{cohens_d.default}
\alias{cohens_d.formula}
\alias{cohens_d.htest}
\title{Cohen's d}
\usage{
cohens_d(x, ...)

\method{cohens_d}{default}(x, y = NULL, paired = FALSE, na.rm = FALSE)

\method{cohens_d}{data.frame}(data, dv, iv, na.rm = FALSE)

\method{cohens_d}{formula}(formula, data, na.rm = FALSE)

\method{cohens_d}{htest}(ttest)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{paired}{a logical indicating whether you want Cohen's d for a paired
sample or two samples.}

\item{na.rm}{logical. Should missing values be removed?}

\item{data}{a data frame containing either the variables in the formula
\code{formula} or the variables specified by \code{dv} and \code{iv}.}

\item{dv}{character indicating the name of the column in \code{data} for the
dependent variable}

\item{iv}{character indicating the name of the column in \code{data} for the
independent variable}

\item{formula}{a formula of the form \code{lhs ~ rhs} where \code{lhs} is a
numeric variable giving the data values and \code{rhs} a factor with two
levels giving the corresponding groups.}

\item{ttest}{an object of class \code{htest} (a call to \code{t_test}).}
}
\description{
Cohen's d
}

