% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa.R
\name{cor_apa}
\alias{cor_apa}
\title{Report Correlation in APA style}
\usage{
cor_apa(x, format = c("text", "markdown", "rmarkdown", "html", "latex",
  "docx", "plotmath"), info = FALSE, print = TRUE)
}
\arguments{
\item{x}{a call to \code{cor.test}}

\item{format}{character string specifying the output format. One of
\code{"text"}, \code{"markdown"}, \code{"rmarkdown"}, \code{html},
\code{"latex"}, \code{"docx"} or \code{"plotmath"}.}

\item{info}{logical indicating whether to print a message on the used test
(default is \code{FALSE})}

\item{print}{logical indicating wheter to print the formatted output via
\code{cat} (\code{TRUE}, default) or return as character string.}
}
\description{
Report Correlation in APA style
}
\examples{
cor_apa(cor.test(~ acrophobia + sens_seek, hquest))

}

