% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apa.R
\name{cor_apa}
\alias{cor_apa}
\title{Report Correlation in APA style}
\usage{
cor_apa(x, format = "text", info = FALSE)
}
\arguments{
\item{x}{a call to \code{cor.test}}

\item{format}{character specifying the output format, one of \code{"text"},
\code{"latex"}, \code{"markdown"}, \code{rmarkdown} or \code{"word"}}

\item{info}{logical indicating whether to print a message on the used test
(default is \code{FALSE})}
}
\description{
Report Correlation in APA style
}
\examples{
ct <- cor.test(runif(20), runif(20))
cor_apa(ct)
cor_apa(ct, format = "latex")
}

