% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplot.R
\name{fplot}
\alias{fplot}
\title{fplot}
\usage{
fplot(.data, formula, geom = c("bar", "line", "boxplot"))
}
\arguments{
\item{.data}{a data frame containing the variables in the formula
\code{formula}.}

\item{formula}{a formula in the form \code{lhs ~ rhs} where \code{lhs} is
the a numeric variable (dependent variable) and \code{rhs} one or more
factors with two or more levels giving the corresponding groups.}

\item{geom}{character indicating the ggplot2 geom used. One of \code{"bar"}
(default), \code{"line"} and \code{"boxplot"}.}
}
\description{
Convenient plotting of means and standard errors or boxplots of factorial
data.
}

