% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{petasq}
\alias{petasq}
\title{Partial Eta Squared}
\usage{
petasq(x, effect)
}
\arguments{
\item{x}{a call to \code{aov}, \code{ez::ezANOVA} or
\code{afex::aov_ez}/\code{afex::aov_car}/\code{afex::aov_4}}

\item{effect}{character string indicating the effect of interest}
}
\description{
Partial Eta Squared
}

