% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recode variables}
\usage{
recode(x, ..., default = NULL, coerce = NULL)
}
\arguments{
\item{x}{numeric or character vector with values to recode.}

\item{...}{recode pairs, e.g. \code{"female" = "f"}.}

\item{default}{numeric or character. All entries in \code{x} not covered by
\code{...} get recoded to \code{default}.}

\item{coerce}{a function that is applied to \code{x} after recoding, e.g.
\code{as.numeric}.}
}
\description{
Recode variables
}
\examples{
recode(c("female", "male", "male", "female"), "female" = "f", "male" = "m")

recode(1:10, "1:5" = "a", "6:10" = "b")

recode(sample(letters[1:6]), "c('a', 'b', 'c')" = "abc",
       "c('d', 'e', 'f')" = "def")
}

