% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats_table.R
\name{stats_table}
\alias{stats_table}
\title{Statistical table}
\usage{
stats_table(.data, iv, ..., sig = FALSE, format = c("default", "html"),
  args = list())
}
\arguments{
\item{.data}{a data frame}

\item{iv}{name of the independent variable}

\item{...}{names of the dependent variables}

\item{sig}{logical indicating whether to show significance with symbols in
the last column}

\item{format}{character string indicating the output format, must be either
\code{"default"} or {"html"}. The latter prints the table as a htmlwidget}

\item{args}{a list with further arguments passed to functions. If the
independent variable has two levels, then \code{args} can contain arguments
passed to \link{t_test}, e.g. \code{var.equal = TRUE}.}
}
\description{
Compiles a table with descriptive and inferential statistics to compare group
characteristics, e.g. control variables of a treatment group and a control
group.
}
\examples{
stats_table(height, iv = group, age:aq_avoi)
stats_table(height, iv = group, age:aq_avoi, sig = TRUE, format = "html",
            args = list(var.equal = TRUE))
}

